+++ controller/con_continents.php
@@ -1,0 +1,4 @@
+<?php
+require_once("/home/cabox/workspace/constants.php");
+require_once(DAL . "/dal_continents.php");
+
