--- controller/con_countries.php
@@ -1,5 +1,0 @@
-<?php
-require_once("/home/cabox/workspace/constants.php");
-require_once(DAL . "/dal_countries.php");
-
-//TODO: Use for AJAX calls
